Create TT Following Auto Copy & Timestamp
Correlated with my video showing how to extract follower counts from TikTok & Instagram using the importxml formula. This code automatically copies the counts extracted from TT & IG and adds it to the next empty row to get a visual log of how things are growing or declining to identify any trends. This code should be companied with a time based trigger to execute at regular intervals